Improve handling of signatures for route handlers
If a handler function had a modified function signature (like a header was removed from it), the generated OpenAPI spec could still end up having the parameter included due to how the handler was copied and the annotation for the request model was added.

This changes how the handler function is copied and how the annotation for the request model is added so it works correctly in the case when the signature of the handler function has been modified.